+++ test/src/player.spec.js
@@ -2845,6 +2845,7 @@ describe('Player', function() {
 
     it('should reset the active text cues', () => {
       player._activeTextCues[0] = {};
+      player._updateTextDisplay = () => {};
       player._resetTextCuesAndReposition();
       let cue = player._activeTextCues[0];
       cue.hasBeenReset.should.equals(true);